// src/equipment/enums/equipment.enum.ts
export enum EquipmentStatus {
    AVAILABLE = 'available',
    RENTED = 'rented',
    MAINTENANCE = 'maintenance',
}

export enum EquipmentCategory {
    LAPTOP = 'laptop',
    DESKTOP = 'desktop',
    SERVER = 'server',
    MONITOR = 'monitor',
    PRINTER = 'printer',
    SCANNER = 'scanner',
    ROUTER = 'router',
    PROJECTOR = 'projector',
    MICROPHONE = 'microphone',
    CAMERA = 'camera',
    OTHER = 'other',
}